滚动类组件布局范围约束提升

+++ frameworks/core/components_ng/pattern/waterflow/layout/top_down/water_flow_layout_info.h
@@ -194,7 +194,7 @@ public:
 
     void InvalidatedOffset() override {};
 
-    float currentOffset_ = 0.0f;
+    double currentOffset_ = 0.0;
     // 0.0f until itemEnd_ is true
     float maxHeight_ = 0.0f;
 
@@ -243,7 +243,7 @@ struct WaterFlowLayoutInfo::ItemInfo {
     }
 
     int32_t crossIdx = 0;
-    float mainOffset = 0.0f;
+    double mainOffset = 0.0;
     float mainSize = 0.0f;
 };
